Combined code w/o new SRAM module
Probably back up your old one too
It doesn't compile because there's more than one of everything in clock_pipe (2x elec4700's, 2x seven_segment, etc. )
6 bit program counter
Turn on switch 9 to start the program, will complete at 50Hz (can comment this out if you want the button back). 
Should have the 4-way cache and 2 processors working as before, but still with the old SRAM module.
cpu_done initialised and passed through both cpus in preparation for when we add the new SRAM
Also the cpu_done variables rely on the instruction set finishing with a "blank" to work, so this is now essential to add.